[ANDROID] Do not record function addresses if value profiling is disabled
Temporary change until a proper fix is merged to upstream.

Don't recording function addresses in __llvm_prf_data when value
profiling is turned off.  This is safe for Android, where we don't use
value profiling.

Functional change: needs -mllvm -enable-value-profiling to work with LTO

The proper fix is to use associated metadata to link a __llvm_prf_data
of a function with the function itself so the linker can discard the
__llvm_prf_data if the function itself is discarded.  See
http://lists.llvm.org/pipermail/llvm-dev/2019-December/137363.html for
more details.

Change-Id: Ifcb04349e71e20ae3ef4bb82ebf8ac064fc59169
Signed-off-by: Jason Edson <jaysonedson@gmail.com>